fix: add support for parsing the document number check digit in more cases
Some recent Belgian and Portuguese ID cards do not include the `<` before the `optional 1` field in the check digit.

I have not found this behavior documented in any official specification so far, but it can be confirmed using documents available online (for example, PRADO). I have also encountered real IDs with this characteristic.